# Improve isascii for short strings 15-30% improvement