This is a tweak that ends up improving `isascii`, particularly for shorted strings.
In #48568 we had made the decision to not use the trick of treating the `codeunits` as `UInt64` this changes that and shows some improvement across the board.

Using the isascii benchmarks in [ndinsmore/StringBenchmarks](https://github.com/ndinsmore/StringBenchmarks) the results are shown below.

There is a version that is another 15-20% faster for short strings, which replaces the 7 byte  epilog `for` loop with the code below, but I could not convince myself that the unsafe load wouldn't cause problems.

```julia
rqword = Base.unsafe_load(pointer(qwords),nqword+1)
rqword >>= (8-epilog_bytes) * 8
return iszero(rqword & UInt64(0x8080808080808080))
```

Before pr:
```julia
julia> results = run(StringBenchmarks.SUITE["isascii"])
4-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "single nonASCII" => 4-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "length 2:64" => Trial(808.355 μs)
	  "length 512:4096" => Trial(43.703 μs)
	  "length 64:512" => Trial(145.182 μs)
	  "length 4096:32768" => Trial(18.661 μs)
  "julia 1.9 source" => 4-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "files" => Trial(207.956 μs)
	  "lines" => Trial(4.876 ms)
	  "files SubString" => Trial(58.269 μs)
	  "lines SubString" => Trial(4.889 ms)
  "Unicode" => 4-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "length 2:64" => Trial(704.782 μs)
	  "length 512:4096" => Trial(39.246 μs)
	  "length 64:512" => Trial(168.937 μs)
	  "length 4096:32768" => Trial(1.587 μs)
  "ASCII" => 4-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "length 2:64" => Trial(716.833 μs)
	  "length 512:4096" => Trial(38.866 μs)
	  "length 64:512" => Trial(131.775 μs)
	  "length 4096:32768" => Trial(30.260 μs)
```

With PR
```julia
julia> results = run(StringBenchmarks.SUITE["isascii"])
4-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "single nonASCII" => 4-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "length 2:64" => Trial(617.714 μs)
	  "length 512:4096" => Trial(49.989 μs)
	  "length 64:512" => Trial(117.014 μs)
	  "length 4096:32768" => Trial(15.793 μs)
  "julia 1.9 source" => 4-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "files" => Trial(152.962 μs)
	  "lines" => Trial(4.453 ms)
	  "files SubString" => Trial(54.206 μs)
	  "lines SubString" => Trial(4.152 ms)
  "Unicode" => 4-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "length 2:64" => Trial(472.361 μs)
	  "length 512:4096" => Trial(34.800 μs)
	  "length 64:512" => Trial(110.567 μs)
	  "length 4096:32768" => Trial(1.310 μs)
  "ASCII" => 4-element BenchmarkTools.BenchmarkGroup:
	  tags: []
	  "length 2:64" => Trial(671.735 μs)
	  "length 512:4096" => Trial(39.734 μs)
	  "length 64:512" => Trial(134.950 μs)
	  "length 4096:32768" => Trial(28.465 μs)
```