@ndinsmore 2023-04-16
So I am out of the office for a week so there will be little update on this during that time , but I have a version that is even better for shorts strings.

The reason that this ends up being faster is that the SIMD reduction operation from a whole register down to a single value is faster for the 64 bit rather than 8 because for a 256 bit register that is a reduction of 4 values rather than 16.