@ndinsmore 2023-04-25
The code has been updated to use include an optimization for lengths < 8 characters as well.  Unfortunately, this means it is a bit less elegant than the version before.

@stevengj We are currently using a chunk size of 1024 bytes which seems optimal for avx256 machines.  This means code is slightly less optimum for avx512 processors as the chunk size is too small for its register size. 

Right now for the full speed SIMD portion which on AVX256 does 4 parallel op the loop process 128 bytes per loop step.  The operations are as follows, assuming n trips through loop:

- 5 ops to initialize the simd registers (vector.ph)
- n * 7 ops for the loop body (vector.body) on my computer this 4x32 bytes = 128 bytes 
- 10 ops to reduce the simd and check the result

Total 1024 byte / 71 cycles => 14.42 bytes/cycle
I think that for avx512 this would be 43 cycles => 23.81 bytes/cycle

This does results in a method that is larger than the previous but I don't know how to quantify that impact.

If we had some basic constant that told us the size of the SIMD registers we could further optimize this, but without that, I think this is about as good as we can get.

@Seelengrab & @matthias314  you help with the other isASCII I would appreciate your feedback as well.