@ndinsmore 2023-04-26
> The one concern I have is with checking the qword preamble - should that check for alignment?

I have tried to see if alignment made a difference and could not get anything to manifest.   I think that since we are copying to the simd register, it doesn't make a difference.