@KristofferC 2023-04-27
My personal opinion is that this is too much and too complicated code for a moderate performance gain. It's fun to hardcore optimize a piece of code but you also have to look at the broader picture. Are there any real life benchmarks where this has a noticeable impact?